Fix pbm with DOMGenerator.
Pbm was that that DOMGenerator was using digest of DOM as an
output cache key and a validty that was provided by it's
instantiator.  This meant that if you had two equivalent docs
, say from some string literals, you could run into cases where
the cached object would be considered invalid. ( Cause k1 == k2
but v1 >= v2 ).

Now each DOMGenerator instance has a unique output key.

For the 512M, 50 thread test this sped things up by
500ms. ( ~7% )