Form Runner proxy portlet: rewrite CSS correctly
New Orbeon-Container header that allows a client to tell the server whether the request originates from a portlet vs. servlet environment. This allows the server to use the proper URL rewriting strategy. For PDF generation, for example, CSS resources are currently rewritten using the "servlet" rewriting strategy, not the "warp" rewriting strategy.