Improved dynamic state serialization
Instead of serializing to an XML document as was the case before, serialize to an immutable DynamicState object which represents a snapshot of a document's dynamic state. That snapshot can be further serialized when needed, for example to disk by Ehcache, or to the client when in client state mode.

Serialization is now mainly external to the objects to serialize, with the exception of SAXStore which already implemented read/writeExternal.

This change required updating wrap-server.xpl, used in unit tests, and xforms-submission.xpl (which underlies oxf:xforms-submission). They used to deal with creating an XML dynamic state representation. This is no longer desirable, so they use oxf:xforms-to-xhtml instead. Serialization of the dynamic state to XML has been updated so that unit tests can still rely on that XML output. A new XForms property, xxforms:no-updates, allows skipping entirely the initial dynamic state serialization. This is only used by oxf:xforms-submission for now.

Some refactoring was done along the way, including moving XFormsInstance and XFormsState to Scala, removing ReadonlyXFormsInstance, etc.