Serialize/deserialize template
The static state was keeping a reference on full update marks but not properly serializing the associated template SAXStore or the marks. SAXStore now serializes/deserializes its marks, and the static state or the dynamic state serialize/deserialize the template when needed.

Added property xxforms:noscript-template="static|dynamic" to specify whether the noscript template must be stored in the static or dynamic state. Default is `static`.

Refactored static state element Metadata class and more.

Unit tests for various aspects of serialization.