Fix #186 "Error inserting two "US Phone Number" in repeat"
Instance elements were not de-indexed. We now build them using the regular child creation mechanism in Model. This is a bit tricky because vals in ModelInstances require the children to be initialized. For now we make them lazy vals to solve the issue but it's not as robust as it could be.