Fix #229 "Fix handling of handlers within repeat"
This also improves the fix for #202, which did not handle xxforms-index-changed and xxforms-nodeset-changed properly.

We have chosen a simple approach: instead of the current hack of redirecting some events from repeat to iteration, we now statically make nested handlers without explicit observers observe not the repeat iteration, but the repeat container itself. This is a compromise which keeps the repeat container/repeat iteration split, still associates the repeat id to the repeat container, but lets nested handlers behave in a more intuitive way. We consider this ok also because at this time we don't dispatch xforms-enabled/disabled to repeat iterations, so there is not much of a point to observe repeat iterations themselves. Via bubbling, listening to the container still allows handling events coming from the iterations, and running them in the correct XPath context.

The resolution algorithm to find the concrete handler was partially broken and is now fixed. The idea is to take the closest handler from the target, not from the observer as was incorrectly the case before. This is a bit harder for the special cases where the observer and the handler are not in the same XBL scope.

Updated existing unit tests for the changed behavior, and added new unit tests, including one for the fr:datatable fr-selection-changed event.