Fix NPE upon error in 2-pass submissions
We didn't pass the XFCD or submission in case of a submission's second pass. This meant that in case of an exception, we would NPE because the document was not there but was used. It might not be nice to hold on to the document between the start of the submission and the completion of its connection, but on the other hand the hold should not last for a very long time.