Fix #539 "Some hidden controls appear"
The issue was that some CSS uses id selectors. Those need to be rewritten. However, in full portlet mode, CSS is served via the servlet. This means that the portlet namespace was not known, and so the ids were not rewritten.

We now pass an extra request parameters so that the CSS rewriter can be made aware of the namespace.