XBL mappings/resources: associate results with properties
Before 28a06f7e1d459f5ddc6174acc9b92aaae96ac0ff, automatic XBL mappings and baseline resources were associated with the static state, and would be re-evaluated when properties were modified, as in that case the entire static state would be recreated.

Now these are global, so we need to do a bit of extra work to re-evaluate them when properties are modified.