Progress on #959 Separate portlet deployment/XForms portlet
- fix forwarding of URL parameters from XForms portlet
- remove OrbeonPortletFilterSample from portlet.xml
- theme-embeddable-examples.xsl doesn't override orbeon class
- recombineQuery adds to existing parameters
- add type parameters (Map<String, String[]> parameters)
- add descriptors for XForms portlet

It's unclear if/when we'll do more work on this. This is the result of spending a bit of time looking at what's working and what's broken.

One thing that was identified was that the implementation of the request wrapper didn't properly override query parameters, so some new code was written to try to do this in a better way.

Another thing was that theme-embeddable-examples.xsl was running, but orbeon class was always overridden.

There is a general issue related to URL rewriting. The XForms engine/Form Runner should probably work in a way similar to the proxy portlet, where URLs are encoded by the servlet and decoded by the portlet. This is not done now, which is a recipe for problems.

Also, the portlet should properly support render and action URLs, as the other portlets do, which means that BufferedPortlet should probably be used.