Processes: separate upload check
It is not right to check for pending uploads at the start of every sub-process. Only some actions need to check that there are pending uploads. Currently:

- save-final
- review
- send
- email
- save
- submit
- workflow-send
- workflow-edit

On the other hand, save-draft for example doesn't check, because you want to save the draft as soon as possible, as things stand.

Also strongly typed the combinators.