Fix issues after e19d4c1f3212b78979ffb18c5c83e8c50bd8f9c2
There were a few issues with how `propertiesStartsWith` was used in conjunction with the new properties in form-runner-properties.xml.

Fixing by:

1. not using `substring` unsafely
2. adding a `matchWildcards` parameter to `propertiesStartsWith`

That's because we really want to make sure, when we request properties starting with `oxf.fr.persistence.provider`, that the properties literally start with that. This excludes e.g. `oxf.fr.persistence.*.autosave`.