Fix #5031: wall not showing with 24-column layout
It is unclear to me why `gridForCell` was returning the nested element
with the class `fr-grid-body`. It makes more sense to return the root
element for the component, with class `xbl-fr-grid`, which we are doing
now, or maybe its child with class `fr-grid`, but since we have code
that assumes that the underlying elements is the root element for the
component, this is what we are returning now.