Fix #5042: allow editing hint for single checkbox
For the label/hint editor to work, the label or hint must be attached to
the control added in the form, not to a control inside that control.
If that happens, the Scala.js code ends up calling
`controlUpdateLabelOrHintOrText()` with the id of contained control,
and we then fail to update the resource.

The bottom line is that the Form Builder label or hint editors won't
work if we're using a custom label or custom hint.