For #4980: fix case where the database is not in 4.0.0 format
There was an impedance mismatch between the result of running the form
in the background and what the export processor was expecting. When
the data comes from the database, then the persistence provider's data
format must be used. But when the data comes from running the form, the
data format is whatever we agree upon. I picked the edge format so as
to avoid extra data transformations.