For #4843: refactor to omit item rather than value
We had the `<xf:copy>` possibly return the empty sequence, which doesn't
make much sense. Instead, when values must be omitted from the itemset,
simply don't return them at all. This also allows, possibly, considering
an item with a label but an empty sequence value to be considered like a
grouping choice.