Keyboard shortcuts: don't use plain "Enter"
The issue is that if we are in the ⌘J dialog and press "Enter", this
closes the dialog but the event is still sent to Form Builder. There is
currently no way for Form Builder to know that the key event was tied to
the dialog and not the main window. We should disable top-level key
listeners when we are in a dialog in the future. For now, changing the
shortcut.